add save result and move to success save result screen

+++ app/src/main/java/ru/eruditeonline/app/data/repository/ResultRepositoryImpl.kt
@@ -51,7 +51,7 @@ class ResultRepositoryImpl @Inject constructor(
     }
 
     override suspend fun saveResult(
-        completeId: Long,
+        completeId: Int,
         name: String,
         surname: String,
         patronymic: String?,
